feat: create analytics custom hook
